+++ neuromation/api/images.py
@@ -131,7 +131,7 @@ class Images(metaclass=NoPublicConstructor):
                 raise AuthorizationError(f"Access denied {remote}") from error
             raise  # pragma: no cover
 
-        await self._docker.images.tag(repo, local)
+        await self._docker.images.tag(repo, str(local))
 
         return local